backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.39
pydantic==2.12.1
python-dotenv==1.1.0
requests==2.32.3
beautifulsoup4==4.13.3
selenium==4.29.0
aiohttp==3.12.15
numpy==2.3.2
scikit-learn==1.7.2
transformers==4.56.2
torch==2.10.0
jinja2==3.1.6
reportlab==4.4.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
redis==5.2.1
celery==5.4.0